+++ CHANGELOG.md
@@ -149,10 +149,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 #### Added
 
-  * `Transaction.withValidStart()`
-  * `Transaction.generate()`
+  * `TransactionId.withValidStart()`
+  * `TransactionId.generate()`
 
 ### Transaction
+
+#### Added
+  * `Transaction.hash()`
+  * `Transaction.signWithOperator()`
 
 #### Removed